Consolidate functions used more than once - gaussian_kernel() - low_rank_eigen()
